mset: Turn iterator into a GObject
While having it as a boxed type that could be placed on the stack helped
the C side of things, binding XapianMSetIterator proved to be a bit more
brittle than I expected. GJS consistently segfaults when handling
iterators, as it seems to be doing multiple copies internally, and
trying to free them multiple times. I honestly think this is a bug in
GJS and/or introspection, but I don't think I'll have the time to
properly follow the rabbit down the hole.

By turning MSetIterator into a GObject everything seems to be working
fine, and the small 'segfault.js' example does not crash any more, so
I'm happy.