stopper: Allow implementing XapianStopper subclasses
In order to implement a XapianStopper sub-class that acts according to
the C++ Xapian::Stopper protocol, we need to use an internal
Xapian::Stopper implementation that calls out to the GObject class
vtable.

We use the ::constructed virtual function of GObject to check if a
XapianStopper subclass, like XapianSimpleStopper, has overridden the
internal Xapian::Stopper pointer, and if it didn't then we can create
our own generic stopper instance, which will then use the instance
pointer to get to the class vtable and call out the sub-class
implementation.

Since XapianQueryParser uses the internal Xapian::Stopper instance, this
has the added benefit of working out of the box with every custom
XapianStopper sub-class.