Always translate the XapianDatabase flags
When creating a new WritableDatabase we take the flags from the Database
base class and we OR them with the ones from the child class. Since we
store the flags as coming from the C API, we need to translate them to
their equivalent C++ values, otherwise they may not match what Xapian
expects.

https://phabricator.endlessm.com/T17251