Split database backend from flags
The option handling in Xapian::Database is a bit of a mess.

The "flags" argument is really made of a bitwise OR of three different
enumerations, with three different semantics:

  - action, an exclusive enumeration
  - backend, an exclusive enumeration
  - flags, a bitmask

In other words, you construct it from:

  - an action value (open, create, create and open, overwrite and open)
  - a backend (glass, chert, stub, in memory)
  - a set of options

Xapian will then mask off the action and backend values, in order to get
to the flags.

It would be easy to just provide the same semantics of Xapian, and give
the app developer a bitmask and construct everything from constant
values; sadly, we already have an Action/Flags split, and Xapian has the
tendency to break ABI between minor releases, which means we'd have to
break API and ABI as well.

Instead, we can fully decouple actions, backend, and flags into separate
properties and enumerations, and construct the flags argument ourselves
when opening the database.

This means:

  - adding a new XapianDatabaseBackend enumeration
  - removing the backend flags from XapianDatabaseFlags
  - adding a XapianDatabase:backend property
  - using xapian_database_get_flags() to build the flags from the
    backend and flags properties

The XapianWritableDatabase can now take the action, and bitwise OR it
with the result of xapian_database_get_flags() safely, as the semantics
of each enumeration are preserved internally.

https://phabricator.endlessm.com/T17251