Fix the declaration of the get_description() method
Methods take an instance pointer as the first argument.

Technically, this is an ABI break; on the other hand, we had a public
symbol with no implementation, and an implementation with the correct
signature that was never made public. Exporting the implementation is
not an ABI break, except that it means dropping the old, unimplemented
symbol because we're exporting a C ABI and we cannot have two symbols
with the same name and different arguments in the same shared object.
Of course, if anybody attempted to use the exported symbol, they would
have found themselves with a linker error, so it's pretty clear nobody
actually tried — not even ourselves, otherwise we would have never
released 1.2 with that glaring issue.

While we reflect on the importance of unit testing and API coverage, and
on the deeper ramifications of ABI stability, I think it's perfectly
acceptable to introduce a small ABI break during a new development
cycle.