Catch exceptions when compacting
Compacting may raise an exception, because the database backend may not
support specific compaction policies.

It's too late to add a GError to the API, so we should emit a critical
warning; this at least allows us to warn the app developer with an error
message, and avoids an abrupt termination.

https://phabricator.endlessm.com/T17839