build: Bind interface_age to version
We want to reset the interface age to 0 every time we are doing unstable
releases, but bind it to the micro version every time we do stable ones.

We only allow adding new API during unstable development cycles, which
means that our interface age does not change during stable cycles.