Wrap more Xapian::Query constructors
This wraps Xapian::Query(Xapian::Query::op, I<T> begin, I<T> end) where I
is an iterator and T is a Xapian::Query, Xapian::Query*, std::string, or
something convertible to a std::string.

Since we don't have that sort of clever overloading in C, we wrap it with
two functions: xapian_query_new_for_terms(XapianQueryOp, char * const *)
which covers the string overload, and
xapian_query_new_for_queries(XapianQueryOp, GSList *) where the value type
of the GSList is XapianQuery *.

For now we ignore the optional termcount window parameter, so the default
value of 0 will use the number of subqueries as the termcount window size.

https://phabricator.endlessm.com/T20373