Add more options to the compact() wrappers
Compacting databases takes a bitmask of options divided into:

 - compacting flags
 - the backend to compact to
 - the compaction level

Since we only ever used a single database backend, and we didn't really
care about compaction levels, we never actually implemented the
functionality required to allow compacting a database using one backend
to a database using another backend.

Now that Xapian supports the Honey backend, though, we are going to need
this functionality, as the Honey backend cannot be used to create
writable databases. In order to use the Honey database backend, you're
supposed to create a WritableDatabase instance using the Glass backend,
and then compact the database using the Honey backend type.

https://phabricator.endlessm.com/T21956